rename movielist class and layout
