+++ app/src/main/AndroidManifest.xml
@@ -16,9 +16,9 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <activity android:name=".MovieList.MovieListActivity"/>
         <activity android:name=".ZukanList.ZukanListActivity"/>
         <activity android:name=".ZukanDetail.ZukanDetailMain"/>
-        <activity android:name=".MovieList.MoveListMainActivity"/>
         <activity android:name=".panframe.SimplePlayerActivity"/>
         <activity android:name=".Quiz.QuizActivity"/>
     </application>
