+++ app/src/main/java/xyz/virtual_diving/projectmainver2/HomeActivity.java
@@ -6,8 +6,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 
+import xyz.virtual_diving.projectmainver2.MovieList.MovieListActivity;
 import xyz.virtual_diving.projectmainver2.ZukanList.ZukanListActivity;
-import xyz.virtual_diving.projectmainver2.MovieList.MoveListMainActivity;
 
 
 public class HomeActivity extends Activity {
@@ -22,7 +22,7 @@ public class HomeActivity extends Activity {
         DM_btm.setOnClickListener(new View.OnClickListener(){
             @Override
             public  void onClick(View v){
-                Intent intent = new Intent(HomeActivity.this, MoveListMainActivity.class); //ダイビングアクティビティに飛ぶ処理
+                Intent intent = new Intent(HomeActivity.this, MovieListActivity.class); //ダイビングアクティビティに飛ぶ処理
                 startActivity(intent);
             }
         });
