+++ app/src/main/res/layout/movielist_listitem.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:background="#E7F1F9"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="horizontal"
+    android:layout_gravity="left"
+    android:weightSum="1">
+    <ImageView
+        android:id="@+id/MovieList_itemimage"
+        android:padding="3dp"
+        android:layout_width="140dp"
+        android:layout_height="140dp"
+        />
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="wrap_content"
+        android:layout_height="137dp"
+        android:orientation="vertical"
+        android:layout_weight="1"
+        android:weightSum="1">
+        <TextView
+            android:textColor="#015DB2"
+            android:text="タイトルa"
+            android:id="@+id/MovieList_titletext"
+            android:textSize="40dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="0.15" />
+        <TextView
+            android:text="本文"
+            android:id="@+id/MovieList_contentstext"
+            android:layout_width="293dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="0.74" />
+        <TextView
+            android:text="再生数"
+            android:id="@+id/MovieList_viewconut"
+            android:textSize="20dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" />
+    </LinearLayout>
+    <ImageView
+        android:paddingTop="50dp"
+        android:layout_width="90dp"
+        android:layout_height="90dp"
+        android:layout_gravity="right"
+        android:src="@drawable/allow_right"
+        />
+</LinearLayout>
