+++ app/src/main/res/layout/simpleplayer_main.xml
